requirements.txt
# Core LangChain packages (latest stable versions as of Dec 2024)
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.0
langchain-core>=0.3.0

# LangSmith for tracing and evaluation
langsmith>=0.1.0

# Vector Store
faiss-cpu>=1.8.0

# OpenAI
openai>=1.50.0

# Utilities
tiktoken>=0.7.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# Note: If you encounter version conflicts, you can create a fresh virtual environment:
# python -m venv venv
# source venv/bin/activate  # On Windows: venv\Scripts\activate
# pip install --upgrade pip
# pip install -r requirements.txt